requirements.txt
# ----- FILE: backend/requirements.txt -----
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
bcrypt==4.0.1
eventlet==0.33.3
geopy==2.3.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2